Swapping Firebase pod (closed source) for FirebaseCommunity pod (open source), as it should end issues with "transitive dependencies on static binaries" that we're seeing (e.g. #48).
Note: due to duplicate symbol issues, if you want to use GeoFire in this way, you'll have to use the FirebaseCommunity pod for other dependencies (Auth, Storage, Messaging), and won't be able to use any other Firebase features.

Once we can resolve the "transitive dependencies" issue in the Firebase pod (merging pods, shipping the pod as dylibs, etc.), we'll switch back to the Firebase pod and the above issue will go away.